+++ src/global_var.cpp
@@ -1,7 +1,7 @@
 #include "global_var.h"
 
 int fd = -1;
-std::atomic<int> x_mmps_setpt(0), theta_mmps_setpt(0);
+std::atomic<double> x_mmps_setpt(0), theta_mmps_setpt(0);
 nav_msgs::Odometry odom;
 int line_sensor = 1000;
